feat: :sparkles: search for hex code array by number
