+++ Program.cs
@@ -63,14 +63,16 @@ app.UseCors("AllowFrontend");
 app.UseAuthentication();
 app.UseAuthorization();
 
-app.MapGet("/", () => "very front end. much display").WithTags("");
-
 app.MapAuthRoutes();
-app.MapDevRoutes();
-app.MapUserRoutes();
 app.MapColourRoutes();
 app.MapCollectionRoutes();
+app.MapDevRoutes();
+app.MapUserRoutes();
+app.MapUtilRoutes();
+
 app.MapIdentityApi<User>()
    .WithTags("Authentication");
 
-app.MapGet("/", () => "JJJ API is running!");
+app.MapGet("/", () => "JJJ API is running!").WithTags("");
+
+app.Run();
